requirements.txt
Django==1.9
djangorestframework==3.3.3
requests==2.10.0
beautifulsoup4==4.6.0
django-filter==0.13.0
celery==3.1.23
